New: added SelectableGridComponent HOC
This PR adds a new HOC component, the SelectableGridComponent.
This HOC only work and was only tests with the FlexTable, it allows you to select and deselect rows with different behavior, with the metaKey, shiftKey and multiple selections.

Example added in the demo application
Unit tests in source/SelectableGridComponent/SelectableGridComponent.test.js
Docs added in the docs/SelectableGridComponent.md