Remove pinnng scrollTop to 0
This is done so that if there are elements which appear above the current element, scrollOffset.top - this._positionFromTop will return a negative number thereby correctly telling the child that it is not being rendered. This allows the child to render only when it becomes necessary to do so as opposed to rendering always.