Fix incorrect inline style of width rendered on the server-side
When rendering on the server-side with the `WindowScroller` component,
the `serverHeight` and `serverWidth` props must be specified,
otherwise, nothing would be rendered.

Unfortunately, most of the time, we can only guess values
for these two props when rendering on the server-side.

There is a `width: ${this.state.width}px` inline style
rendered on the server-side on the inner `Grid` component,
and the value of `this.state.width` is derived from `serverWidth`.

Since the value of `serverWidth` is not likely equal to window's width,
`this.state.width` must be fixed on the client-side when the component mounted
and then triggering a re-rendering for the `WindowScroll` component
with the correct value.

But the `this.state.width` is initialized to a CORRECT value on the client-side,
that make the call to `updatePosition()` in `componentDidMount()` useless.
Hence the INCORRECT inline style would never be fixed
unless the user resizes the browser window.

This fix initialized `this.state.width` with the value of`serverWidth`
on the client-side if `serverWidth` is specified,
so that the call to `updatePosition()` in `componentDidMount()`
can actually trigger a re-rendering.