Better handle calculation of cell size
Holding a cache of cell sizes can have a memory hit in case of very
large lists, especially if the user scrolls to a deep page. This
mitigates the issue in case the CellSize is a numeric value. In such
a case, instead of calculating the size of all cells up to the wanted
index and caching the results, we simply multiply the index by the height.